Docs for public members.
